fixed errors in formula to calculate output

+++ java/com/example/milindmahajan/mortgagecalculator/CalculationHelper.java
@@ -8,31 +8,31 @@ import static java.lang.Math.pow;
  */
 public class CalculationHelper {
 
-    public static HashMap<String, String> calculateMortgageFrom(Float homeValue, Float downPay, float terms, float apr, Float propertyTax) {
+    public static HashMap<String, String> calculateMortgageFrom(Double homeValue, Double downPay, Double terms, Double apr, Double propertyTax) {
 
         HashMap<String, String> calculationMap = new HashMap<String, String>();
 
-        float actualLoan = homeValue - downPay;
-        float totalTerms = terms * 12;
+        Double actualLoan = homeValue - downPay;
+        Double totalTerms = terms * 12;
 
-        float totalPropertyTax = homeValue * propertyTax/100 * terms;
+        Double totalPropertyTax = homeValue * propertyTax/100 * terms;
         calculationMap.put("propertyTax", String.format("%.2f", totalPropertyTax));
 
-        float monthlyPropertyTaxTemp = totalPropertyTax/totalTerms;
-        float monthlyPropertyTax = Float.parseFloat(String.format("%.2f", monthlyPropertyTaxTemp));
+        Double monthlyPropertyTaxTemp = totalPropertyTax/totalTerms;
+        Double monthlyPropertyTax = Double.parseDouble(String.format("%.2f", monthlyPropertyTaxTemp));
 
-        float monthlyInterestRate = apr/12;
+        Double monthlyInterestRate = apr/12;
         monthlyInterestRate /= 100;
 
-        double temp = pow((1+monthlyInterestRate), totalTerms);
+        Double temp = pow((1+monthlyInterestRate), totalTerms);
 
-        double monthlyMortgagePaymentTemp = (actualLoan * monthlyInterestRate * temp)/(temp - 1);
-        double monthlyMortgagePayment = Double.parseDouble(String.format("%.2f", monthlyMortgagePaymentTemp));
+        Double monthlyMortgagePaymentTemp = (actualLoan * monthlyInterestRate * temp)/(temp - 1);
+        Double monthlyMortgagePayment = Double.parseDouble(String.format("%.2f", monthlyMortgagePaymentTemp));
 
-        double monthlyPayment = monthlyMortgagePayment + monthlyPropertyTax;
+        Double monthlyPayment = monthlyMortgagePayment + monthlyPropertyTax;
         calculationMap.put("monthlyPayment", String.format("%.2f", monthlyPayment));
 
-        double totalInterestPaid = (monthlyPayment * totalTerms) - (actualLoan + downPay);
+        Double totalInterestPaid = (monthlyPayment * totalTerms) - (actualLoan + downPay);
         calculationMap.put("totalInterest", String.format("%.2f", totalInterestPaid));
 
         return calculationMap;
